*Follow up r1334

+++ src/map/skill.h
@@ -368,7 +368,7 @@ void ext_skill_unit_onplace(struct skill_unit *src, struct block_list *bl, unsig
 int skill_unit_ondamaged(struct skill_unit *src,struct block_list *bl,int64 damage,unsigned int tick);
 
 int skill_castfix(struct block_list *bl, uint16 skill_id, uint16 skill_lv);
-int skill_castfix_sc(struct block_list *bl, int time);
+int skill_castfix_sc(struct block_list *bl, double time);
 #ifdef RENEWAL_CAST
 int skill_vfcastfix(struct block_list *bl, double time, uint16 skill_id, uint16 skill_lv);
 #endif
